backend/requirements.txt
alembic==1.13.3
asyncpg==0.30.0
bcrypt==4.2.0
email-validator==2.2.0
fastapi==0.115.4
greenlet==3.1.1
passlib[bcrypt]==1.7.4
pydantic==2.9.2
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.17
psycopg[binary]==3.2.3
redis==5.1.1
SQLAlchemy==2.0.36
uvicorn[standard]==0.32.0
